initial rework of motion infra to support stateful skills
