+++ control-board/src/motion/mod.rs
@@ -1,6 +1,8 @@
+pub mod body_controller;
 pub mod constant_gain_kalman_filter;
+pub mod control_context;
 pub mod pid;
-pub mod robot_controller;
 pub mod robot_model;
+pub mod skills;
 
 pub mod params;
